backend/requirements.txt
langchain
python-dotenv
langchain-openai
langchain-core
langchain-community
bs4
faiss-cpu
from_root
langchain_groq
transformers
wikipedia
arxiv

IPython

flask
flask-cors
pymysql
openpyxl
flask_sqlalchemy
flask_migrate
flask_login
flask_wtf
flask_bootstrap
flask_mail
flask_babel
plotly
langchain
python-dotenv
langchain-openai
langchain-community
bs4
faiss-cpu
from_root
langchain_groq
langchain-community
llama-parse
fastembed
chromadb
python-dotenv
langchain-groq
chainlit
fastembed
# unstructured[md]
